📝 Add Jekyll documentation
This change attempts to move us away from our `README.md`, which we've
outgrown, into a series of Jekyll-compatible Markdown files, which
focus on better segmentation of the documentation, and having more
headings that can be linked to.

The aim of this documentation is to be hosted on Github Pages.