🥅 Remove agent truthiness checks
We recently made [a change][1], which added checks for if `agent` was
provided to certain backend methods. This added unofficial support for
a falsy `agent`, which we don't really want to support.

This change removes these checks, and fixes the tests that fail because
they pass in a `null` agent.

Note that we discussed adding a formal assertion, but this pattern
didn't fix with the existing methods, which all operate under the
assumption that its arguments are present.

[1]: https://github.com/share/sharedb/pull/446