🔨 Fix example script builds on Windows
Running the example `build` commands on windows doesn't work, because
`mkdir` [fails with the wrong slash type][1].

However, `browserify` will already [ensure the directory exists][2], so
we don't even need this command.

This change removes our call to `mkdir`, and also simplifies the call
to `browserify`, since `node_modules` is already on the PATH when run
with `npm run`.

[1]: https://stackoverflow.com/a/52863346/4003671
[2]: https://github.com/browserify/browserify/blob/0ec6e80ec48b67513718a392a6d09bd5569967d4/bin/cmd.js#L64