🥅 Improve presence error message
At the moment, if a consumer tries to listen for presence on a type that
does not support presence, they'll get an error along the lines of:

```
TypeError: this._doc.type.transformPresence is not a function
    at LocalDocPresence._transformAgainstOp
```

This can be a bit confusing for consumers who assume their type supports
presence.

This change wraps that method in a function that will check if the type
supports presence. If not, then we throw a more descriptive error,
explaining that the type does not support presence.