🥅 Normalize `json0` `lm`
We recently [added a shim][1] to let consumers bump to a breaking
version of `json0`.

This shim normalized the `json0` path to allow older, less strict ops to
be applied using the stricter version of `json0`.

However, that change also included [stricter checking for `lm`][2].
This change updates our shim to also allow for these older ops. We also
rename our internal methods to be more reflective of the fact that we're
normalizing ops (rather than just their paths).

[1]: https://github.com/share/sharedb/pull/496
[2]: https://github.com/ottypes/json0/pull/40/files#diff-188db14d4925d79e94b18d68782b7df264ff95d95a26812b5b2c07c90f5640afR222-R223