🚨 Ignore docs when linting
For some reason, the build has started linting our `/docs/` folder. It's
possible that this was an `eslint` bug, and that it should have always
been linted, since we just use our `.gitignore` file (and the JS in our
docs folder is ignored by its own, nested `.gitignore` file).

This change bumps our version of `eslint`, and adds the `/docs`/ folder
separately to be ignored by `eslint`.