✨ Add `receivePresence` middleware hook
At the moment, we have a [`sendPresence`][1] middleware hook, which is
triggered just before a presence update is about to be sent to a client.

This existing hook is useful for determining if a client should receive
an update or not on the server.

However, it doesn't let us *stop* clients from broadcasting to other
clients, which we may want to do for authentication reasons, for
example.

This change adds a new `receivePresence` middleware hook, which is
triggered when the server receives a presence message from a client,
before the presence has been transformed against any ops.

Passing an error to this middleware hook will return an error message to
the broadcasting client (as opposed to `sendPresence`, which will send
an error down to any subscribed clients).

[1]: https://github.com/share/sharedb/blob/07f8bd98e5cdae37e925d4c0c709a1c3cbbca5e0/lib/agent.js#L816