🔒 Prevent sending `sendPresence` middleware errors to subscribers
At the moment, if an error is passed in the `sendPresence` middleware,
this error is broadcast to any subscribed listeners.

Consumers may wish to use the `sendPresence` middleware to filter
messages being sent to particular clients (eg based on authentication or
other criteria). In these cases, no information at all should be
broadcast to subscribers, since this might be something of a security
risk - a malicious client can tell there's activity on a channel, even
if they don't know exactly what that activity is.

This change deprecates the current error behaviour, and adds a new flag
to opt-in to the new behaviour of emitting errors on `Backend` instead
of sending them down to clients. If consumers haven't actively opted in
for this behaviour, they will get a deprecation warning.