🐛 Prevent local presence transformation when presence created late
This change fixes a presence edge case:

 1. Register a `doc.on('op')` listener which submits presence
 2. Create the `LocalPresence` _after_ registering this listener
 3. Submit an op
 4. The presence being submitted is transformed against the submitted op
    when it shouldn't be

This happens because the `LocalPresence` handles its transformation
logic in the `'op'` event. If the `LocalPresence` is created after
another handler that submits presence, then its listener is bound late,
and is only triggered after the handler that submits the op.

The result is that the `LocalPresence` will try to transform the
presence, even though the presence is already consistent with the doc's
state.

We could in theory try to prepend the `LocalPresence` listener, but we'd
still be vulnerable to the same bug if consumers _also_ prepend
listeners, and we just move the problem to another place.

Instead, this change fixes the issue by tracking the state of `doc.data`
when the presence is submitted.

However, since `doc.version` is only incremented when an op is
[acknowledged][1], we can't use this to track our document state. To
address this, this change adds a new `_dataStateVersion` internal
variable, which is incremented every time `doc.data` is changed. This
allows us to compare the state of our local `doc.data` over time and
we can use this to check if our presence needs transformation or not.

[1]: https://github.com/share/sharedb/blob/a1a33852531432baeb3cb948ca36500bdbc60fa4/lib/client/doc.js#L947-L948