✨ Allow middleware to mutate submitted ops
This change adds a new `request.$fixup()` method to the `SubmitRequest`
class, which allows middleware to mutate ops in the `apply` middleware.

Motivation
----------
There are times when we want the server to make changes to ops submitted
by a client. There may be many reasons for wanting to do this, including
distrusting a client (eg setting timestamp, or auth information);
keeping sensitive business logic out of client-side code; etc.

The only ways of doing this currently are:

 1. Committing the client's op, and then submitting an amendment op from
    the server; or
 2. Rejecting the client's op, and submitting the full amended op from
    the server

Option 1 works, but possibly commits badly-formed, or untrustworthy data
to the database, and broadcasts it to remote clients. It also makes it
possible to fetch historic snapshots of this untrustworthy in-between
state, which undermines the trustworthiness of our data.

Option 2 also works, and might work with invertible types through a
"soft" rollback, but can lose pending ops if we're not careful, or if
the type isn't invertible.

Approach
--------
This change adds a `$fixup()` helper function to the `SubmitRequest`
object, named after the `git` action for performing similar actions.

The idea of the `$fixup()` function is to act as a middleware-friendly
`submitRequest()`.

Any ops submitted through `$fixup()` will be:

 - composed with the `request.op`, so downstream middleware can work on
   the fixed-up op
 - sent back to the original client, and applied
 - help to avoid infinite middleware loops, since we don't trigger any
   nested submission calls

Note that in order to handle the case of a client disconnecting before
getting their fixup, we now also store the fixup ops on the op `m`
metadata field, and fetch this when getting ops to the current snapshot.

Limitations
-----------
This feature will only work:

 - on types which implement the optional `.compose()` function
 - in the `apply` middleware, where the snapshot's `type` is available,
   but before the op has actually been applied to the snapshot
 - on creation or edit ops (not deletion)