⚡️ Deduplicate `fetch()` network requests
At the moment, if a client calls `doc.fetch()` with a fetch request
already in progress, the client will send a second request, which
probably isn't needed (or wanted).

This change prevents duplicate requests from being sent when there's
already an inflight fetch request, and adds an internal `force` option,
which can be used to force a duplicate request to be sent. This is
currently used internally for `_hardRollback()`, which forcibly unsets
the `doc.version`, and therefore requires a separate `fetch()` call,
since fetch requests are [handled differently][1] depending on the
presence of `version`.

[1]: https://github.com/share/sharedb/blob/21750e320f6865ee14620d6bf9610cb7ebc5f18e/lib/agent.js#L607