✨ Store client protocol on `Agent`
When developing new features, it will be useful to be able to compare
server and client protocol versions.

For example, if a server is ahead of an old client, the server could
gracefully degrade performance for clients that don't support the new
functionality.

This change stores the client's protocol on the `Agent` so we can
perform these checks.

We also add a `checkAtLeast()` protocol helper function, and store the
current protocol in a constant to avoid magic numbers.