🐛 Fix `closed` check when handling `querySubscribe()`
The `Agent` attempts to check if it's been closed when getting the
response for a query subscribe.

However, it incorrectly tries to access `this` inside a `function`,
which doesn't give the correct value, and in `strict mode`, will
actually result in an error, since `this` will be `undefined`.

This change adds a test for this case and fixes it.